doc: update AUTHORS
Thinking it is probably better to run this frequently and do small quick
timely updates to AUTHORS rather than waiting months like we usually do
and having a lot of little issues to sort out. If this works well, maybe
it can be a scheduled GitHub Action or something the bot does for us.

PR-URL: https://github.com/nodejs/node/pull/39217
Reviewed-By: Michaël Zasso <targos@protonmail.com>
Reviewed-By: Richard Lau <rlau@redhat.com>