tools: remove Babel from license-builder.sh
Babel is now installed as a dependency in our ESLint installation. The
current license-builder.sh will fail to find the license for Babel.

The license is sill present in the repository/code base/file system. So
as with all other dependencies that are installed in
tools/node_modules/eslint/node_modules and have an MIT license, we are
still complying with the license. The inclusion in our own LICENSE was
to clearly signal that things under tools/node_modules/@babel were not
covered by the Node.js license. Because it is now under
tools/node_modules/eslint, that "we aren't the license-holders of this
code" aspect of things is already covered by the inclusion of the ESLint
license.

PR-URL: https://github.com/nodejs/node/pull/41049
Reviewed-By: Michaël Zasso <targos@protonmail.com>
Reviewed-By: Antoine du Hamel <duhamelantoine1995@gmail.com>
Reviewed-By: Luigi Pinca <luigipinca@gmail.com>
Reviewed-By: Tobias Nießen <tniessen@tnie.de>
Reviewed-By: James M Snell <jasnell@gmail.com>